Improve the gitignore and delete all PDFs
We don't want any build artifacts inside the repository...
We could use CI+CD to automatically release new versions.

The PDF directory was responsible for ~60% of the total size of a fresh
clone (repository incl. the working tree). But the size of the
repository itself (--bare) is luckily still at 3.9M, due to delta
(and normal) compression.

$ du -s --si .
13M	.
$ du -s --si pdf
7.7M	pdf
$ du -s --si .git
3.9M	.git