Makefile: Make pdftk optional (only required for the map)
Since Inkscape is already optional there is no guarantee that we
actually build the map and we can make pdftk optional as well.
As both tools are present in the CI/CD pipeline the final result will
always contain the map.

Note: Inkscape was not actually optional since all letters include
stadtplan.pdf and the build would fail if stadtplan.pdf is not present.
As a workaround I'm using pdftex to generate an empty PDF document.