Fix failing build
Inkscape is capable of directly converting SVG to PDF files.
Since at least version > 0.92.1, the current Makefile does not work as
expected and probably produces a PDF in $(OUTDIR)/stadtplan.tmp.
Also remove all references to PDFtk because we don't need it anymore
(since Inkscape is converting the SVG to a PDF directly).