Fix the Nix CI build with TeX Live 2019
Since TeX Live 2019 the KOMA-Script class `scrartcl` seems to cause
errors that make the build fail [0]. I didn't manage to figure out what
exactly is going wrong, but since `scrartcl` tends to cause such
problems it seems best to switch to the core document class `article`
anyway (causes way fewer problems in general - at least IMO).

This obviously causes a few visual differences (the headings look a bit
less "nice" and the indentation is a bit larger for both paragraphs and
footnotes), but the result doesn't change significantly (without a
direct comparison it normally shouldn't be noticeable).
I switched to Babel for proper hyphenation (avoids too long lines), and
replaced a "\pagebreak" with "\newpage" to fix the position of the
footnote.

[0]: https://travis-ci.org/fsi-tue/anfibrief/builds/653176728